requirements.txt
# Gemfury repo
--extra-index-url https://repo.fury.io/${MC_FURY_AUTH}/imakecodes/

colorama
django
django-extensions
django-json-widget
django-minio-storage
django-mysql
django-storages
djangorestframework
djangorestframework-api-key
GitPython
mysqlclient
pendulum
pillow
requests
sentry-sdk
sorl-thumbnail
